requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beanie==1.26.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jwt==1.3.1
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.5.0
orjson==3.10.5
packaging==24.1
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
toml==0.10.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
